[add]: added dms_kafka_instances
