+++ .gitignore
@@ -23,3 +23,5 @@ terraform-provider-sbercloud
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+
+.idea/
