+++ sbercloud/acceptance/dms/resource_sbercloud_dms_kafka_instance_test.go
@@ -1,35 +1,36 @@
-package sbercloud
+package dms
 
 import (
 	"fmt"
+	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/acceptance"
 	"testing"
 
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
-	"github.com/chnsz/golangsdk/openstack/dms/v2/rabbitmq/instances"
+	"github.com/chnsz/golangsdk/openstack/dms/v2/kafka/instances"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
-func TestAccDmsRabbitmqInstances_basic(t *testing.T) {
+func TestAccDmsKafkaInstances_basic(t *testing.T) {
 	var instance instances.Instance
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
 	updateName := rName + "update"
-	resourceName := "sbercloud_dms_rabbitmq_instance.test"
+	resourceName := "sbercloud_dms_kafka_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckDmsRabbitmqInstanceDestroy,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckDmsKafkaInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDmsRabbitmqInstance_basic(rName),
+				Config: testAccDmsKafkaInstance_basic(rName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
+					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", rName),
-					resource.TestCheckResourceAttr(resourceName, "engine", "rabbitmq"),
+					resource.TestCheckResourceAttr(resourceName, "engine", "kafka"),
 				),
 			},
 			{
@@ -38,65 +39,66 @@ func TestAccDmsRabbitmqInstances_basic(t *testing.T) {
 				ImportStateVerify: true,
 				ImportStateVerifyIgnore: []string{
 					"password",
+					"manager_password",
 					"used_storage_space",
 				},
 			},
 			{
-				Config: testAccDmsRabbitmqInstance_update(rName, updateName),
+				Config: testAccDmsKafkaInstance_update(rName, updateName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
+					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", updateName),
-					resource.TestCheckResourceAttr(resourceName, "description", "rabbitmq test update"),
+					resource.TestCheckResourceAttr(resourceName, "description", "kafka test update"),
 				),
 			},
 		},
 	})
 }
 
-func TestAccDmsRabbitmqInstances_withEpsId(t *testing.T) {
+func TestAccDmsKafkaInstances_withEpsId(t *testing.T) {
 	var instance instances.Instance
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
-	resourceName := "sbercloud_dms_rabbitmq_instance.test"
+	resourceName := "sbercloud_dms_kafka_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheckEpsID(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckDmsRabbitmqInstanceDestroy,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckDmsKafkaInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDmsRabbitmqInstance_withEpsId(rName),
+				Config: testAccDmsKafkaInstance_withEpsId(rName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
+					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", rName),
-					resource.TestCheckResourceAttr(resourceName, "engine", "rabbitmq"),
-					resource.TestCheckResourceAttr(resourceName, "enterprise_project_id", SBC_ENTERPRISE_PROJECT_ID_TEST),
+					resource.TestCheckResourceAttr(resourceName, "engine", "kafka"),
+					resource.TestCheckResourceAttr(resourceName, "enterprise_project_id", acceptance.SBC_ENTERPRISE_PROJECT_ID_TEST),
 				),
 			},
 		},
 	})
 }
 
-func testAccCheckDmsRabbitmqInstanceDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*config.Config)
-	dmsClient, err := config.DmsV2Client(SBC_REGION_NAME)
+func testAccCheckDmsKafkaInstanceDestroy(s *terraform.State) error {
+	config := acceptance.TestAccProvider.Meta().(*config.Config)
+	dmsClient, err := config.DmsV2Client(acceptance.SBC_REGION_NAME)
 	if err != nil {
 		return fmtp.Errorf("Error creating SberCloud dms instance client: %s", err)
 	}
 
 	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "sbercloud_dms_rabbitmq_instance" {
+		if rs.Type != "sbercloud_dms_kafka_instance" {
 			continue
 		}
 
 		_, err := instances.Get(dmsClient, rs.Primary.ID).Extract()
 		if err == nil {
-			return fmtp.Errorf("The Dms rabbitmq instance still exists.")
+			return fmtp.Errorf("The Dms kafka instance still exists.")
 		}
 	}
 	return nil
 }
 
-func testAccCheckDmsRabbitmqInstanceExists(n string, instance instances.Instance) resource.TestCheckFunc {
+func testAccCheckDmsKafkaInstanceExists(n string, instance instances.Instance) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
@@ -107,26 +109,26 @@ func testAccCheckDmsRabbitmqInstanceExists(n string, instance instances.Instance
 			return fmtp.Errorf("No ID is set")
 		}
 
-		config := testAccProvider.Meta().(*config.Config)
-		dmsClient, err := config.DmsV2Client(SBC_REGION_NAME)
+		config := acceptance.TestAccProvider.Meta().(*config.Config)
+		dmsClient, err := config.DmsV2Client(acceptance.SBC_REGION_NAME)
 		if err != nil {
 			return fmtp.Errorf("Error creating SberCloud dms instance client: %s", err)
 		}
 
 		v, err := instances.Get(dmsClient, rs.Primary.ID).Extract()
 		if err != nil {
-			return fmtp.Errorf("Error getting SberCloud dms rabbitmq instance: %s, err: %s", rs.Primary.ID, err)
+			return fmtp.Errorf("Error getting SberCloud dms kafka instance: %s, err: %s", rs.Primary.ID, err)
 		}
 
 		if v.InstanceID != rs.Primary.ID {
-			return fmtp.Errorf("The Dms rabbitmq instance not found.")
+			return fmtp.Errorf("The Dms kafka instance not found.")
 		}
 		instance = *v
 		return nil
 	}
 }
 
-func testAccDmsRabbitmqInstance_Base(rName string) string {
+func testAccDmsKafkaInstance_Base(rName string) string {
 	return fmt.Sprintf(`
 data "sbercloud_dms_az" "test" {}
 
@@ -139,78 +141,87 @@ data "sbercloud_vpc_subnet" "test" {
 }
 
 data "sbercloud_dms_product" "test" {
-  engine        = "rabbitmq"
+  engine        = "kafka"
   instance_type = "cluster"
-  version       = "3.7.17"
+  version       = "2.3.0"
 }
 
 resource "sbercloud_networking_secgroup" "test" {
   name        = "%s"
-  description = "secgroup for rabbitmq"
+  description = "secgroup for kafka"
 }
 `, rName)
 }
 
-func testAccDmsRabbitmqInstance_basic(rName string) string {
+func testAccDmsKafkaInstance_basic(rName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_rabbitmq_instance" "test" {
+resource "sbercloud_dms_kafka_instance" "test" {
   name              = "%s"
-  description       = "rabbitmq test"
+  description       = "kafka test"
   access_user       = "user"
-  password          = "Rabbitmqtest@123"
+  password          = "Kafkatest@123"
   vpc_id            = data.sbercloud_vpc.test.id
   network_id        = data.sbercloud_vpc_subnet.test.id
   security_group_id = sbercloud_networking_secgroup.test.id
   available_zones   = [data.sbercloud_dms_az.test.id]
   product_id        = data.sbercloud_dms_product.test.id
   engine_version    = data.sbercloud_dms_product.test.version
+  bandwidth         = data.sbercloud_dms_product.test.bandwidth
   storage_space     = data.sbercloud_dms_product.test.storage
   storage_spec_code = data.sbercloud_dms_product.test.storage_spec_code
+  manager_user      = "kafka-user"
+  manager_password  = "Kafkatest@123"
 }
-`, testAccDmsRabbitmqInstance_Base(rName), rName)
+`, testAccDmsKafkaInstance_Base(rName), rName)
 }
 
-func testAccDmsRabbitmqInstance_update(rName, updateName string) string {
+func testAccDmsKafkaInstance_update(rName, updateName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_rabbitmq_instance" "test" {
+resource "sbercloud_dms_kafka_instance" "test" {
   name              = "%s"
-  description       = "rabbitmq test update"
+  description       = "kafka test update"
   access_user       = "user"
-  password          = "Rabbitmqtest@123"
+  password          = "Kafkatest@123"
   vpc_id            = data.sbercloud_vpc.test.id
   network_id        = data.sbercloud_vpc_subnet.test.id
   security_group_id = sbercloud_networking_secgroup.test.id
   available_zones   = [data.sbercloud_dms_az.test.id]
   product_id        = data.sbercloud_dms_product.test.id
   engine_version    = data.sbercloud_dms_product.test.version
+  bandwidth         = data.sbercloud_dms_product.test.bandwidth
   storage_space     = data.sbercloud_dms_product.test.storage
   storage_spec_code = data.sbercloud_dms_product.test.storage_spec_code
+  manager_user      = "kafka-user"
+  manager_password  = "Kafkatest@123"
 }
-`, testAccDmsRabbitmqInstance_Base(rName), updateName)
+`, testAccDmsKafkaInstance_Base(rName), updateName)
 }
 
-func testAccDmsRabbitmqInstance_withEpsId(rName string) string {
+func testAccDmsKafkaInstance_withEpsId(rName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_rabbitmq_instance" "test" {
+resource "sbercloud_dms_kafka_instance" "test" {
   name                  = "%s"
-  description           = "rabbitmq test"
+  description           = "kafka test"
   access_user           = "user"
-  password              = "Rabbitmqtest@123"
+  password              = "Kafkatest@123"
   vpc_id                = data.sbercloud_vpc.test.id
   network_id            = data.sbercloud_vpc_subnet.test.id
   security_group_id     = sbercloud_networking_secgroup.test.id
   available_zones       = [data.sbercloud_dms_az.test.id]
   product_id            = data.sbercloud_dms_product.test.id
   engine_version        = data.sbercloud_dms_product.test.version
+  bandwidth             = data.sbercloud_dms_product.test.bandwidth
   storage_space         = data.sbercloud_dms_product.test.storage
   storage_spec_code     = data.sbercloud_dms_product.test.storage_spec_code
+  manager_user          = "kafka-user"
+  manager_password      = "Kafkatest@123"
   enterprise_project_id = "%s"
 }
-`, testAccDmsRabbitmqInstance_Base(rName), rName, SBC_ENTERPRISE_PROJECT_ID_TEST)
+`, testAccDmsKafkaInstance_Base(rName), rName, acceptance.SBC_ENTERPRISE_PROJECT_ID_TEST)
 }
