+++ sbercloud/acceptance/dms/resource_sbercloud_dms_rabbitmq_instance_test.go
@@ -1,35 +1,36 @@
-package sbercloud
+package dms
 
 import (
 	"fmt"
+	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/acceptance"
 	"testing"
 
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
-	"github.com/chnsz/golangsdk/openstack/dms/v2/kafka/instances"
+	"github.com/chnsz/golangsdk/openstack/dms/v2/rabbitmq/instances"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
-func TestAccDmsKafkaInstances_basic(t *testing.T) {
+func TestAccDmsRabbitmqInstances_basic(t *testing.T) {
 	var instance instances.Instance
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
 	updateName := rName + "update"
-	resourceName := "sbercloud_dms_kafka_instance.test"
+	resourceName := "sbercloud_dms_rabbitmq_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckDmsKafkaInstanceDestroy,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckDmsRabbitmqInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDmsKafkaInstance_basic(rName),
+				Config: testAccDmsRabbitmqInstance_basic(rName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
+					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", rName),
-					resource.TestCheckResourceAttr(resourceName, "engine", "kafka"),
+					resource.TestCheckResourceAttr(resourceName, "engine", "rabbitmq"),
 				),
 			},
 			{
@@ -38,66 +39,65 @@ func TestAccDmsKafkaInstances_basic(t *testing.T) {
 				ImportStateVerify: true,
 				ImportStateVerifyIgnore: []string{
 					"password",
-					"manager_password",
 					"used_storage_space",
 				},
 			},
 			{
-				Config: testAccDmsKafkaInstance_update(rName, updateName),
+				Config: testAccDmsRabbitmqInstance_update(rName, updateName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
+					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", updateName),
-					resource.TestCheckResourceAttr(resourceName, "description", "kafka test update"),
+					resource.TestCheckResourceAttr(resourceName, "description", "rabbitmq test update"),
 				),
 			},
 		},
 	})
 }
 
-func TestAccDmsKafkaInstances_withEpsId(t *testing.T) {
+func TestAccDmsRabbitmqInstances_withEpsId(t *testing.T) {
 	var instance instances.Instance
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
-	resourceName := "sbercloud_dms_kafka_instance.test"
+	resourceName := "sbercloud_dms_rabbitmq_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheckEpsID(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckDmsKafkaInstanceDestroy,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckDmsRabbitmqInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDmsKafkaInstance_withEpsId(rName),
+				Config: testAccDmsRabbitmqInstance_withEpsId(rName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDmsKafkaInstanceExists(resourceName, instance),
+					testAccCheckDmsRabbitmqInstanceExists(resourceName, instance),
 					resource.TestCheckResourceAttr(resourceName, "name", rName),
-					resource.TestCheckResourceAttr(resourceName, "engine", "kafka"),
-					resource.TestCheckResourceAttr(resourceName, "enterprise_project_id", SBC_ENTERPRISE_PROJECT_ID_TEST),
+					resource.TestCheckResourceAttr(resourceName, "engine", "rabbitmq"),
+					resource.TestCheckResourceAttr(resourceName, "enterprise_project_id", acceptance.SBC_ENTERPRISE_PROJECT_ID_TEST),
 				),
 			},
 		},
 	})
 }
 
-func testAccCheckDmsKafkaInstanceDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*config.Config)
-	dmsClient, err := config.DmsV2Client(SBC_REGION_NAME)
+func testAccCheckDmsRabbitmqInstanceDestroy(s *terraform.State) error {
+	config := acceptance.TestAccProvider.Meta().(*config.Config)
+	dmsClient, err := config.DmsV2Client(acceptance.SBC_REGION_NAME)
 	if err != nil {
 		return fmtp.Errorf("Error creating SberCloud dms instance client: %s", err)
 	}
 
 	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "sbercloud_dms_kafka_instance" {
+		if rs.Type != "sbercloud_dms_rabbitmq_instance" {
 			continue
 		}
 
 		_, err := instances.Get(dmsClient, rs.Primary.ID).Extract()
 		if err == nil {
-			return fmtp.Errorf("The Dms kafka instance still exists.")
+			return fmtp.Errorf("The Dms rabbitmq instance still exists.")
 		}
 	}
 	return nil
 }
 
-func testAccCheckDmsKafkaInstanceExists(n string, instance instances.Instance) resource.TestCheckFunc {
+func testAccCheckDmsRabbitmqInstanceExists(n string, instance instances.Instance) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
@@ -108,26 +108,26 @@ func testAccCheckDmsKafkaInstanceExists(n string, instance instances.Instance) r
 			return fmtp.Errorf("No ID is set")
 		}
 
-		config := testAccProvider.Meta().(*config.Config)
-		dmsClient, err := config.DmsV2Client(SBC_REGION_NAME)
+		config := acceptance.TestAccProvider.Meta().(*config.Config)
+		dmsClient, err := config.DmsV2Client(acceptance.SBC_REGION_NAME)
 		if err != nil {
 			return fmtp.Errorf("Error creating SberCloud dms instance client: %s", err)
 		}
 
 		v, err := instances.Get(dmsClient, rs.Primary.ID).Extract()
 		if err != nil {
-			return fmtp.Errorf("Error getting SberCloud dms kafka instance: %s, err: %s", rs.Primary.ID, err)
+			return fmtp.Errorf("Error getting SberCloud dms rabbitmq instance: %s, err: %s", rs.Primary.ID, err)
 		}
 
 		if v.InstanceID != rs.Primary.ID {
-			return fmtp.Errorf("The Dms kafka instance not found.")
+			return fmtp.Errorf("The Dms rabbitmq instance not found.")
 		}
 		instance = *v
 		return nil
 	}
 }
 
-func testAccDmsKafkaInstance_Base(rName string) string {
+func testAccDmsRabbitmqInstance_Base(rName string) string {
 	return fmt.Sprintf(`
 data "sbercloud_dms_az" "test" {}
 
@@ -140,87 +140,78 @@ data "sbercloud_vpc_subnet" "test" {
 }
 
 data "sbercloud_dms_product" "test" {
-  engine        = "kafka"
+  engine        = "rabbitmq"
   instance_type = "cluster"
-  version       = "2.3.0"
+  version       = "3.7.17"
 }
 
 resource "sbercloud_networking_secgroup" "test" {
   name        = "%s"
-  description = "secgroup for kafka"
+  description = "secgroup for rabbitmq"
 }
 `, rName)
 }
 
-func testAccDmsKafkaInstance_basic(rName string) string {
+func testAccDmsRabbitmqInstance_basic(rName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_kafka_instance" "test" {
+resource "sbercloud_dms_rabbitmq_instance" "test" {
   name              = "%s"
-  description       = "kafka test"
+  description       = "rabbitmq test"
   access_user       = "user"
-  password          = "Kafkatest@123"
+  password          = "Rabbitmqtest@123"
   vpc_id            = data.sbercloud_vpc.test.id
   network_id        = data.sbercloud_vpc_subnet.test.id
   security_group_id = sbercloud_networking_secgroup.test.id
   available_zones   = [data.sbercloud_dms_az.test.id]
   product_id        = data.sbercloud_dms_product.test.id
   engine_version    = data.sbercloud_dms_product.test.version
-  bandwidth         = data.sbercloud_dms_product.test.bandwidth
   storage_space     = data.sbercloud_dms_product.test.storage
   storage_spec_code = data.sbercloud_dms_product.test.storage_spec_code
-  manager_user      = "kafka-user"
-  manager_password  = "Kafkatest@123"
 }
-`, testAccDmsKafkaInstance_Base(rName), rName)
+`, testAccDmsRabbitmqInstance_Base(rName), rName)
 }
 
-func testAccDmsKafkaInstance_update(rName, updateName string) string {
+func testAccDmsRabbitmqInstance_update(rName, updateName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_kafka_instance" "test" {
+resource "sbercloud_dms_rabbitmq_instance" "test" {
   name              = "%s"
-  description       = "kafka test update"
+  description       = "rabbitmq test update"
   access_user       = "user"
-  password          = "Kafkatest@123"
+  password          = "Rabbitmqtest@123"
   vpc_id            = data.sbercloud_vpc.test.id
   network_id        = data.sbercloud_vpc_subnet.test.id
   security_group_id = sbercloud_networking_secgroup.test.id
   available_zones   = [data.sbercloud_dms_az.test.id]
   product_id        = data.sbercloud_dms_product.test.id
   engine_version    = data.sbercloud_dms_product.test.version
-  bandwidth         = data.sbercloud_dms_product.test.bandwidth
   storage_space     = data.sbercloud_dms_product.test.storage
   storage_spec_code = data.sbercloud_dms_product.test.storage_spec_code
-  manager_user      = "kafka-user"
-  manager_password  = "Kafkatest@123"
 }
-`, testAccDmsKafkaInstance_Base(rName), updateName)
+`, testAccDmsRabbitmqInstance_Base(rName), updateName)
 }
 
-func testAccDmsKafkaInstance_withEpsId(rName string) string {
+func testAccDmsRabbitmqInstance_withEpsId(rName string) string {
 	return fmt.Sprintf(`
 %s
 
-resource "sbercloud_dms_kafka_instance" "test" {
+resource "sbercloud_dms_rabbitmq_instance" "test" {
   name                  = "%s"
-  description           = "kafka test"
+  description           = "rabbitmq test"
   access_user           = "user"
-  password              = "Kafkatest@123"
+  password              = "Rabbitmqtest@123"
   vpc_id                = data.sbercloud_vpc.test.id
   network_id            = data.sbercloud_vpc_subnet.test.id
   security_group_id     = sbercloud_networking_secgroup.test.id
   available_zones       = [data.sbercloud_dms_az.test.id]
   product_id            = data.sbercloud_dms_product.test.id
   engine_version        = data.sbercloud_dms_product.test.version
-  bandwidth             = data.sbercloud_dms_product.test.bandwidth
   storage_space         = data.sbercloud_dms_product.test.storage
   storage_spec_code     = data.sbercloud_dms_product.test.storage_spec_code
-  manager_user          = "kafka-user"
-  manager_password      = "Kafkatest@123"
   enterprise_project_id = "%s"
 }
-`, testAccDmsKafkaInstance_Base(rName), rName, SBC_ENTERPRISE_PROJECT_ID_TEST)
+`, testAccDmsRabbitmqInstance_Base(rName), rName, acceptance.SBC_ENTERPRISE_PROJECT_ID_TEST)
 }
