+++ sbercloud/iam/credentials.csv
@@ -1,0 +1,2 @@
+﻿User ID,Access Key ID,Secret Access Key
+e7271c2a497f4fc4a7094d364f040bc3,8HSSJWMLVFW5ID2CRELD,cimjiWI744PJsnz8vtBmkvEh1Mk9LCXkJbYdZfnP
